Added right-checking for delete, and filtering in list of robots

+++ app/static_files/js/app.js
@@ -1,7 +1,9 @@
 var app = new Vue({
       el: '#app',
       data: {
-        robots: []
+        robots: [],
+        rightCanDeleteRobot: true,
+        filtertext: ''
       },
       mounted: function() {
         this.loadRobots();
@@ -12,13 +14,32 @@ var app = new Vue({
           console.log("Vue discovered that robot was added!");
         })
       },
+      // watch: {
+      //   filtertext: function(newfilter) {
+      //     this.filtertext = newfilter;
+      //   }
+      // },
       methods: {
+        getRightsTooltip: function(right, message) {
+          //Used to get an tooltip text for the given right.
+          //If user has the given right, then return the given message
+          //Else if the user does not have the right, then return a standard denied-text.
+          return (right ? message : 'You do not have enough rights!');
+        },
         reverseList: function() {
           this.robots.reverse();
         },
+        filteredRobots: function() {
+          var value = this.filtertext;
+          return this.robots.filter(function(robot){
+            //Compares each of the elements in the robot list, with the filtervalue.
+            return robot.name.toLowerCase().indexOf(value) !== -1
+            || robot.owner.toLowerCase().indexOf(value) !== -1;
+          });
+        },
         deleteRobot: function(robot) {
           //Call on the backend REST API to delete the robot
-          if (confirm('Are you SURE you want to delete ' + robot.name + '?')) {
+          if (this.rightCanDeleteRobot && confirm('Are you SURE you want to delete ' + robot.name + '?')) {
 
               this.$http.delete('/api/robots/' + robot.name).then(function(response) {
                 //alert('ok');
